added register page

+++ base/views.py
@@ -2,7 +2,9 @@ from django.shortcuts import render,redirect
 from django.contrib import messages
 from django.http import HttpResponse
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from django.db.models  import Q
 from .models import Room, Topic
 from . forms import RoomForm
@@ -10,6 +12,10 @@ from . forms import RoomForm
 # Create your views here.
 
 def loginPage(request):
+    page = 'login'  # Set the page context to 'login'
+    
+    if request.user.is_authenticated:
+        return redirect('home')
     
     if request.method == 'POST':
         username = request.POST.get('username')
@@ -17,19 +23,40 @@ def loginPage(request):
 
         try:
             user = User.objects.get(username=username)
-        except:
-            messages.error(request, "user does not exist.")
+        except User.DoesNotExist:
+            messages.error(request, "User does not exist.")
 
         user = authenticate(request, username=username, password=password)
 
         if user is not None:
-            login(request,user)
+            login(request, user)
             return redirect('home')
         else:
-            messages.error(request, "username and passwords not exist")
+            messages.error(request, "Username or password is incorrect.")
 
-    
-    return render(request, 'base/login_register.html')
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+
+def registerPage(request):
+    page = 'register'  # Set the page context to 'register'
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            form.save()
+            messages.success(request, "Account created successfully!")
+            return redirect('login')
+        else:
+            messages.error(request, "An error occurred during registration.")
+
+    context = {'form': form, 'page': page}
+    return render(request, 'base/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
 
 
 def home(request):
@@ -53,6 +80,8 @@ def room(request,pk):
     return render(request, 'base/room.html',context)
 
 
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -65,9 +94,14 @@ def createRoom(request):
     return render(request, 'base/room_form.html',context)
 
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id = pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here")
+
     if request.method == 'POST':
 
         form = RoomForm(request.POST, instance=room)
@@ -80,11 +114,16 @@ def updateRoom(request,pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
 
     room = Room.objects.get(id=pk)
 
+    if request.user != room.host:
+            return HttpResponse("You are not allowed here")
+    
     if request.method == 'POST':
+        
         room.delete()
         return redirect('home')
